Added channel.h file for convinience. Modified exit and kill functions in proc.c for destroying all channels when a process exits or is killed.
