+++ kernel/defs.h
@@ -10,6 +10,7 @@ struct stat;
 struct superblock;
 struct channel;
 
+
 // bio.c
 void            binit(void);
 struct buf*     bread(uint, uint);
@@ -193,4 +194,4 @@ void            virtio_disk_intr(void);
 void            channelinit(void);
 int             channel_create(void);
 int             channel_put(int, int);
-int             channel_destory(int);
+int             channel_destroy(int);
