+++ kernel/syschannel.c
@@ -12,6 +12,7 @@
 #include "file.h"
 #include "fcntl.h"
 
+
 extern int channel_create(void);
 extern int channel_put(int id, int data);
 extern int channel_take(int cd, uint64 data);
